requirements.txt
streamlit
Pillow
torch
ultralytics
numpy
matplotlib
streamlit-drawable-canvas
google-cloud-aiplatform
protobuf
vertexai
google-auth